feat: comprehensive code quality improvements
This PR implements a systematic 7-phase code quality improvement plan
addressing security, maintainability, and Go best practices.

## Changes Summary

### Phase 1: Pre-commit Hooks & Formatting
- Add .pre-commit-config.yaml with gofumpt, goimports, golangci-lint
- Update Makefile to use gofumpt and add install-tools target

### Phase 2: Fix staticcheck Issues (23 fixes)
- Replace deprecated ioutil with os.ReadFile/WriteFile
- Create typed context keys to avoid string collisions
- Convert for-select loops to for-range idiom
- Remove redundant nil checks

### Phase 3: Fix Security Issues (38+ fixes)
- Add configurable host key verification for SSH
- Implement HTTP server timeouts (ReadHeaderTimeout, etc.)
- Enforce minimum TLS 1.2 in Prometheus client
- Create internal/fileperms package with type-safe permission constants
- Fix file permissions across auth, logging, and preferences

### Phase 4: Standardize Error Handling
- Add ViewError() constructor to errors package
- Migrate app.go to use S9sError with context
- Migrate DAO layer (30+ sites) to use domain-specific error constructors
- All errors now include stack traces and structured context

### Phase 5: Standardize Logging
- Initialize structured logger in main.go
- Replace all log.Printf with zerolog structured logging
- Add logger field to S9s struct for consistent logging

### Phase 6: Code Cleanup & Refactoring
- Remove 135 lines of dead code from app.go
- Refactor app.go from 1,018 lines to 307 lines (70% reduction)
- Split into focused files:
  - app_commands.go (96 lines)
  - app_keyboard.go (150 lines)
  - app_modals.go (170 lines)
  - app_plugins.go (123 lines)
  - app_ui.go (64 lines)
  - app_views.go (163 lines)

### Phase 7: Additional Linters
- Enable misspell, bodyclose, errorlint, wastedassign
- Fix 3 wastedassign issues

### Errorlint Fixes (10 fixes)
- Use errors.As() instead of type assertions on errors
- Use %w format verb for proper error wrapping
- Use errors.Is() instead of direct error comparisons
- Ensures compatibility with Go 1.13+ error wrapping

## Impact

- **Security**: 38 vulnerabilities addressed
- **Maintainability**: 70% reduction in app.go size
- **Error Handling**: Consistent S9sError with stack traces
- **Code Quality**: Passing all enabled linters
- **Logging**: Structured logging throughout

## Files Changed
- 27 modified files
- 11 new files (app refactor + fileperms package)
- Net: +1479 lines, -909 lines

## Testing
- ✅ All tests passing
- ✅ Build successful
- ✅ Linting passes (0 issues)

Closes #<issue-number>
