+++ plugins/observability/data/secrets/.key_hint
@@ -1,6 +1,6 @@
 {
-  "created_at": "2026-01-21T00:36:19Z",
+  "created_at": "2026-01-23T07:30:34Z",
   "env_var": "OBSERVABILITY_MASTER_KEY",
-  "key_hash": "aff47bf8886d34e9",
+  "key_hash": "d462d29a5905a1ae",
   "source": "environment"
 }
