+++ plugins/observability/initialization/data/secrets/.key_hint
@@ -1,6 +1,6 @@
 {
-  "created_at": "2025-08-10T16:13:07+01:00",
+  "created_at": "2026-01-23T07:30:33Z",
   "env_var": "OBSERVABILITY_MASTER_KEY",
-  "key_hash": "1ecd01349e8ac066",
+  "key_hash": "177947cc41b02460",
   "source": "environment"
 }
